requirements.txt
Flask==2.3.2
Flask_Cors==3.0.10
langchain==0.0.201
pinecone_client==2.2.1
